feature: hop; ui impl, mgr changes, pass 1

+++ app/src/main/java/com/celzero/bravedns/wireguard/WgHopManager.kt
@@ -2,12 +2,16 @@ package com.celzero.bravedns.wireguard
 
 import Logger
 import Logger.LOG_TAG_PROXY
+import com.celzero.bravedns.RethinkDnsApplication.Companion.DEBUG
 import com.celzero.bravedns.database.WgHopMap
 import com.celzero.bravedns.database.WgHopMapRepository
 import com.celzero.bravedns.service.ProxyManager.ID_WG_BASE
 import com.celzero.bravedns.service.VpnController
 import com.celzero.bravedns.service.WireguardManager
 import com.celzero.bravedns.service.WireguardManager.INVALID_CONF_ID
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import java.util.concurrent.CopyOnWriteArrayList
@@ -18,7 +22,15 @@ object WgHopManager: KoinComponent {
     private var maps: CopyOnWriteArrayList<WgHopMap> = CopyOnWriteArrayList()
     private const val TAG = "WgHopMgr"
 
+    init {
+        io { load() }
+    }
+
     suspend fun load(): Int {
+        if (maps.isNotEmpty()) {
+            Logger.i(LOG_TAG_PROXY, "$TAG reload hop: ${maps.size}")
+        }
+        maps.clear()
         maps = CopyOnWriteArrayList(db.getAll())
         printMaps()
         Logger.i(LOG_TAG_PROXY, "$TAG load complete: ${maps.size}")
@@ -28,20 +40,23 @@ object WgHopManager: KoinComponent {
     private suspend fun add(map: WgHopMap): Pair<Boolean, String> {
         val srcConfig = WireguardManager.getConfigById(toId(map.src))
         val viaConfig = WireguardManager.getConfigById(toId(map.via))
-        val canRoute = canRoute(map.src, map.via)
+
+        val isAlreadyMapped = maps.any { it.src == map.src && it.via == map.via }
+        if (isAlreadyMapped) {
+            Logger.i(LOG_TAG_PROXY, "$TAG add: already mapped ${map.src} -> ${map.via}")
+            return Pair(false, "Already mapped ${map.src} -> ${map.via}")
+        }
+
+        val canRoute = canRoute(map.src)
         if (!canRoute) {
             Logger.i(LOG_TAG_PROXY, "$TAG add: Can't route $map.src -> $map.via")
             return Pair(false, "Can't route ${map.src} -> ${map.via}")
         }
+
         if (srcConfig != null && viaConfig != null) {
             Logger.i(LOG_TAG_PROXY, "$TAG add: ${map.src}(${srcConfig.getName()}) -> ${map.via}(${viaConfig.getName()})")
-            val c = VpnController.createWgHop(srcConfig, viaConfig)
-            Logger.i(LOG_TAG_PROXY, "$TAG add: ${c.first}, ${c.second}")
-            val temp = VpnController.via(ID_WG_BASE + srcConfig.getId())
-            Logger.i(LOG_TAG_PROXY, "$TAG via for ${ID_WG_BASE + srcConfig.getId()}: $temp")
-            val temp1 = VpnController.via(ID_WG_BASE + viaConfig.getId())
-            Logger.i(LOG_TAG_PROXY, "$TAG via for ${ID_WG_BASE + viaConfig.getId()}: $temp1")
-            val hop = VpnController.createWgHop(srcConfig, viaConfig)
+            val hop = VpnController.createWgHop(map.src, map.via)
+            Logger.i(LOG_TAG_PROXY, "$TAG add: ${map.src} -> ${map.via}, res: ${hop.first}, ${hop.second}")
             if (hop.first) {
                 map.isActive = true
                 db.insert(map)
@@ -51,37 +66,33 @@ object WgHopManager: KoinComponent {
                 return Pair(false, hop.second)
             }
         } else {
-            Logger.i(LOG_TAG_PROXY, "$TAG add: Invalid config")
+            Logger.i(LOG_TAG_PROXY, "$TAG add: invalid config")
             return Pair(false, "Invalid config")
         }
     }
 
-    private suspend fun delete(map: WgHopMap) {
-        db.delete(map)
-        maps.remove(map)
-        val srcConfig = WireguardManager.getConfigById(toId(map.src))
-        val viaConfig = WireguardManager.getConfigById(toId(map.via))
-        if (srcConfig != null && viaConfig != null) {
-            val hop = VpnController.createWgHop(srcConfig, null)
-            Logger.i(LOG_TAG_PROXY, "$TAG delete: ${hop.first}, ${hop.second}")
+    private suspend fun delete(map: WgHopMap): Pair<Boolean, String> {
+        var res = Pair(false, "Map not found")
+        val rmv = maps.remove(map)
+        if (!rmv) {
+            Logger.i(LOG_TAG_PROXY, "$TAG delete: map not found")
+            return res
         }
-        Logger.i(LOG_TAG_PROXY, "$TAG delete: ${map.src} -> ${map.via}")
+        res = VpnController.createWgHop(map.src, "")
+        db.delete(map)
+        Logger.i(LOG_TAG_PROXY, "$TAG delete: ${map.src} -> ${map.via}, res: $res")
+        return Pair(rmv, "Removed ${map.src} -> ${map.via}")
     }
 
-    suspend fun deleteBySrc(src: String): Pair<Boolean, String> {
+    suspend fun removeHop(srcId: Int, viaId: Int): Pair<Boolean, String> {
+        val src = ID_WG_BASE + srcId
+        val via = ID_WG_BASE + viaId
         var res = Pair(false, "Map not found")
-        val map = maps.find { it.src == src }
-        Logger.d(LOG_TAG_PROXY, "$TAG deleteById: $src")
-        if (map != null) {
-            db.deleteById(map.id)
-            maps.remove(map)
-            val srcConfig = WireguardManager.getConfigById(toId(map.src))
-            val viaConfig = WireguardManager.getConfigById(toId(map.via))
-            if (srcConfig != null && viaConfig != null) {
-                res = VpnController.createWgHop(srcConfig, null)
-                Logger.i(LOG_TAG_PROXY, "$TAG delete: ${res.first}, ${res.second}")
-            }
-            Logger.i(LOG_TAG_PROXY, "$TAG delete: ${map.src} -> ${map.via}")
+        val isAvailableInMap = maps.find { it.src == src && it.via == via }
+        if (DEBUG) printMaps()
+        Logger.v(LOG_TAG_PROXY, "$TAG removeHop: $src")
+        if (isAvailableInMap != null) {
+            res = delete(isAvailableInMap)
         } else {
             Logger.i(LOG_TAG_PROXY, "$TAG delete: map not found")
         }
@@ -91,33 +102,61 @@ object WgHopManager: KoinComponent {
     suspend fun hop(src: Int, via: Int):  Pair<Boolean, String> {
         val srcId = ID_WG_BASE + src
         val viaId = ID_WG_BASE + via
-        val map = maps.find { it.src == srcId && it.via == viaId }
+        // see if the src is mapped with some other via, if so remove it
+        val srcMaps = getMapBySrc(srcId)
+        if (srcMaps.isNotEmpty()) {
+            srcMaps.forEach {
+                if (it.via != viaId && it.via.isNotEmpty()) {
+                    removeHop(src, toId(it.via))
+                }
+            }
+        }
         Logger.d(LOG_TAG_PROXY, "$TAG hop init: $srcId -> $viaId")
-        return if (map != null) {
-            Logger.i(LOG_TAG_PROXY, "$TAG hop: ${map.src} -> ${map.via}")
-            add(map)
+        return if (srcId == viaId) {
+            Pair(false, "Can't hop to self")
         } else {
-            Logger.i(LOG_TAG_PROXY, "$TAG hop: map null, create one with $src -> $via")
-            add(WgHopMap(0, srcId, viaId, false, ""))
+            val map = getMap(srcId, viaId)
+            if (map != null) {
+                Pair(true, "Already hopped")
+            } else {
+                add(WgHopMap(0, srcId, viaId, true, ""))
+            }
         }
     }
 
-    suspend fun getHopableWgs(src: String): List<Config> {
+    suspend fun getVia(src: Int): String {
+        val srcId = ID_WG_BASE + src
+        return getVia(srcId)
+    }
+
+    suspend fun getVia(src: String): String {
+        val map = maps.find { it.src == src && it.isActive }
+        return map?.via ?: ""
+    }
+
+    suspend fun getHopableWgs(src: Int): List<Config> {
+        val srcId = ID_WG_BASE + src
+        val possibleHops: MutableSet<Config> = mutableSetOf()
+        // add the via of this src to the list even if it is not active
+        val via = maps.find { it.src == srcId }
+        if (via != null) {
+            val viaConfig = WireguardManager.getConfigById(toId(via.via))
+            if (viaConfig != null) {
+                possibleHops.add(viaConfig)
+            }
+        }
         // get the list of active wgs
         val activeWgs = WireguardManager.getActiveConfigs()
         Logger.d(LOG_TAG_PROXY, "$TAG getHopableWgs activeWgs: $activeWgs")
-        // see if there is any map with src as via, if so return empty list
-        val viaWgs = maps.filter { it.via == src }
-        if (viaWgs.isNotEmpty()) {
-            Logger.d(LOG_TAG_PROXY, "$TAG getHopableWgs viaWgs: $viaWgs")
-            return emptyList()
-        }
-        Logger.d(LOG_TAG_PROXY, "$TAG getHopableWgs viaWgs: $viaWgs")
-        val possibleHops = activeWgs.filter { wg ->
-            wg.getId() != toId(src) && canRoute(src, wg.getId().toString())
-        }
-        Logger.d(LOG_TAG_PROXY, "$TAG getHopableWgs possibleHops: ${possibleHops.map { it.getName() }}")
-        return possibleHops
+        possibleHops.addAll(activeWgs.filter { wg ->
+            wg.getId() != toId(srcId) && canRoute(srcId)
+        })
+        Logger.i(LOG_TAG_PROXY, "$TAG getHopableWgs possibleHops: ${possibleHops.map { it.getName() }}")
+        return possibleHops.toList() // return the immutable list
+    }
+
+    fun isAlreadyVia(id: String): Boolean {
+        return maps.any { it.via == id }
     }
 
     fun getMaps(): List<WgHopMap> {
@@ -128,15 +167,23 @@ object WgHopManager: KoinComponent {
         return maps.find { it.src == src && it.via == via }
     }
 
+    fun getAllVia(): List<String> {
+        // get all the via from the maps
+        return maps.map { it.via }
+    }
+
     fun getMapBySrc(src: String): List<WgHopMap> {
-        return maps.filter { it.src == src }
+        val filtered = maps.filter { it.src == src }
+        Logger.d(LOG_TAG_PROXY, "$TAG getMapBySrc: $src, filtered: $filtered")
+        printMaps()
+        return filtered
     }
 
     fun getMapByVia(via: String): List<WgHopMap> {
         return maps.filter { it.via == via }
     }
 
-    fun canRoute(src: String, via: String): Boolean {
+    fun canRoute(src: String): Boolean {
         // if the src is via for any other map, then it can't route
         val srcMaps = getMapBySrc(src)
         val can = srcMaps.none { it.via == src }
@@ -149,13 +196,17 @@ object WgHopManager: KoinComponent {
             val configId = id.substring(ID_WG_BASE.length)
             configId.toIntOrNull() ?: INVALID_CONF_ID
         } catch (e: Exception) {
-            Logger.i(LOG_TAG_PROXY, "$TAG err converting string id to int: $id")
+            Logger.i(LOG_TAG_PROXY, "$TAG err converting string id to int: $id, ${e.message}")
             INVALID_CONF_ID
         }
     }
 
     fun printMaps() {
         Logger.v(LOG_TAG_PROXY, "$TAG printMaps: $maps")
+    }
+
+    private fun io(f: suspend () -> Unit) {
+        CoroutineScope(Dispatchers.IO).launch { f() }
     }
 
 }
